refactor: move gnl to utils/
