+++ utils/get_next_line.c
@@ -1,27 +1,5 @@
 #include "get_next_line.h"
 
-static char	*strappend(char *dst, const char *src, size_t l)
-{
-	char	*rtn;
-	size_t	i;
-	size_t	j;
-	size_t	k;
-
-	rtn = (char *)malloc(sizeof(char) * (ft_strlen(dst) + l + 1));
-	if (!rtn)
-		return (NULL);
-	i = 0;
-	j = 0;
-	while (dst[j])
-		rtn[i++] = dst[j++];
-	k = 0;
-	while (k < l)
-		rtn[i++] = src[k++];
-	rtn[i] = '\0';
-	free(dst);
-	return (rtn);
-}
-
 static t_gnl_status	process_buffer(char *buff, char **line)
 {
 	const char	*nl_position = ft_strchr(buff, '\n');
