+++ utils/utils.h
@@ -1,0 +1,11 @@
+#ifndef UTILS_H
+# define UTILS_H
+
+# include <stdlib.h>
+# include <unistd.h>
+
+# include "../libft/libft.h"
+
+char	*strappend(char *dst, const char *src, size_t l);
+
+#endif
